+++ src/zfit/core/binnedpdf.py
@@ -247,7 +247,7 @@ class BaseBinnedPDF(
             return self.space
         if isinstance(x, PlottableHistogram) and not isinstance(x, ZfitBinnedData):
             x = BinnedData.from_hist(x)
-        if not isinstance(x, (ZfitBinnedData, ZfitSpace, ZfitUnbinnedData)):
+        if not isinstance(x, ZfitBinnedData | ZfitSpace | ZfitUnbinnedData):
             # TODO: should we allow spaces? Or what?
             try:
                 x = Data.from_tensor(obs=self.obs, tensor=x)
@@ -907,7 +907,7 @@ class BaseBinnedPDF(
                 raise BreakingAPIChangeError(msg)
             if fixed_params is True:
                 fixed_params = None  # default behavior is to catch all anyway
-            if isinstance(fixed_params, (list, tuple)):
+            if isinstance(fixed_params, list | tuple):
                 fixed_params = {param.name: znp.asarray(param) for param in fixed_params}
             params = fixed_params